Split text visitors into individual files
