+++ vnext/Microsoft.ReactNative/Views/Text/TextHighlighterVisitor.h
@@ -1,0 +1,38 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include <Utils/TextTransform.h>
+#include <Views/Text/TextVisitor.h>
+#include <stack>
+
+namespace Microsoft::ReactNative {
+
+class TextHighlighterVisitor : public TextVisitor {
+  using Super = TextVisitor;
+  using Color = std::optional<winrt::Windows::UI::Color>;
+
+ public:
+  TextHighlighterVisitor(Color foregroundColor, Color backgroundColor) : Super() {
+    m_foregroundColors.push(foregroundColor);
+    m_backgroundColors.push(backgroundColor);
+  }
+
+  std::vector<xaml::Documents::TextHighlighter> highlighters{};
+
+ protected:
+  void VisitRawText(ShadowNodeBase *node) override;
+
+  void VisitVirtualText(ShadowNodeBase *node) override;
+
+ private:
+  int m_startIndex{0};
+  std::stack<Color> m_foregroundColors;
+  std::stack<Color> m_backgroundColors;
+
+  bool RequiresTextHighlighter(Color foregroundColor, Color backgroundColor);
+  static bool Push(std::stack<Color> &stack, Color color);
+};
+
+} // namespace Microsoft::ReactNative
